+++ frontend/app/src/TaskAway/GlobalContext/initialState.js
@@ -1,0 +1,6 @@
+export const initialState = {
+    todoList: [{}],
+	title: '',
+	desc: '',
+    error: null,
+};
